Move Location and ApiError models into mzutils
