+++ src/cmake/sources.cmake
@@ -113,8 +113,6 @@ target_sources(mozillavpn-sources INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/models/devicemodel.h
     ${CMAKE_CURRENT_SOURCE_DIR}/models/keys.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/models/keys.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/models/location.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/models/location.h
     ${CMAKE_CURRENT_SOURCE_DIR}/models/recentconnections.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/models/recentconnections.h
     ${CMAKE_CURRENT_SOURCE_DIR}/models/recommendedlocationmodel.cpp
