+++ src/utils/models/apierror.cpp
